@@ -4,7 +4,8 @@ version: 2.1.0
 homepage: https://github.com/mauriciotogneri/dalocale
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  path: ^1.7.0
+  collection: ^1.15.0-nullsafety.4
+  path: ^1.8.0
